disco, listar y crear

+++ src/main/java/cl/playground/disconorte/repository/DiscoRepository.java
@@ -1,10 +1,13 @@
 package cl.playground.disconorte.repository;
 
 import cl.playground.disconorte.entity.Disco;
+import org.hibernate.query.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.lang.Integer;
+
+import java.util.List;
 
 @Repository
-public interface DiscoRepository extends JpaRepository<Disco, Integer> {
+public interface DiscoRepository extends JpaRepository<Disco, Long> {
 }
